feature(#491): Add cache location metrics
- Mainly for depth2, though I may eventually find a use for it once I have
  multiple depth1 static caches.

+semver: feature

+++ include/fordyca/ds/arena_map.hpp
@@ -81,8 +81,8 @@ class arena_map : public er::client<arena_map>,
  public:
   explicit arena_map(const struct params::arena::arena_map_params* params);
 
-  /* arena metrics */
-  bool has_robot(size_t i, size_t j) const override;
+  /* robot occupancy metrics */
+  bool has_robot(uint i, uint j) const override;
 
   /**
    * @brief Reset the # of caches that have been removed on a single timestep
